requirements.txt
# Stock Trading Advisor - Python Dependencies
# Install with: pip install -r requirements.txt

# ===================================================================
# CORE ESSENTIALS (Required to run the system)
# ===================================================================
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28

# ===================================================================
# ENVIRONMENT & CONFIG
# ===================================================================
python-dotenv>=1.0.0

# ===================================================================
# OPTIONAL ENHANCEMENTS (Install as needed)
# ===================================================================

# Additional Data Sources (optional)
# alpha-vantage>=2.3.1
# polygon-api-client>=1.12.0
# newsapi-python>=0.2.7

# Database (optional - SQLite is built-in)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0  # For PostgreSQL
# alembic>=1.11.0

# Notifications (optional)
# twilio>=8.5.0  # For SMS
# sendgrid>=6.10.0  # For email via SendGrid

# Technical Analysis Libraries (optional)
# ta-lib>=0.4.28  # Requires system installation
# pandas-ta>=0.3.14b0

# Machine Learning (for future phases)
# scikit-learn>=1.3.0
# xgboost>=1.7.0
# lightgbm>=4.0.0

# Deep Learning (for future phases)
# tensorflow>=2.13.0
# torch>=2.0.0

# Time Series Analysis (for future phases)
# statsmodels>=0.14.0
# prophet>=1.1.4

# NLP & Sentiment (for future phases)
# nltk>=3.8.0
# transformers>=4.30.0
# textblob>=0.17.0

# API & Web Utilities
# requests>=2.31.0
# beautifulsoup4>=4.12.0

# Visualization (optional)
# matplotlib>=3.7.0
# seaborn>=0.12.0
# plotly>=5.15.0

# Utilities
# pyyaml>=6.0
# schedule>=1.2.0
# tqdm>=4.65.0

# Testing & Development (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# jupyter>=1.0.0
# ipython>=8.14.0
# black>=23.7.0
# flake8>=6.0.0

# ===================================================================
# MINIMAL INSTALL (Just to get started):
# pip install pandas numpy yfinance python-dotenv
# ===================================================================